+++ ui/dev/src/store/index.js
@@ -1,30 +1,16 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 
-import users from './api/users'
-
-// import example from './module-example'
+import users from './modules/users'
 
 Vue.use(Vuex)
 
-/*
- * If not building with SSR mode, you can
- * directly export the Store instantiation;
- *
- * The function below can be async too; either use
- * async/await or return a Promise which resolves
- * with the Store instance.
- */
-
-export default async function (/* { ssrContext } */) {
+export default function () {
   const Store = new Vuex.Store({
     modules: {
       users
-      // example
     },
 
-    // enable strict mode (adds overhead!)
-    // for dev mode only
     strict: process.env.DEBUGGING
   })