+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed `#q-app` background-color 
+## 1.7.0 - 2021-04-13
+
+### Added
+- Added component tag `q-form` in `QasDialog` if prop `useForm` is `true`
+- Added trigger validate `q-form` if prop `useForm` and emit value
+- Added example page in `ui/dev` with name `DialogForm.vue`
 
 ## 1.6.0 - 2021-04-06